+++ pages/extUI/dateformat/dateformat.vue
@@ -5,14 +5,14 @@
 		</uni-card>
 		<uni-section title="Basic usage" type="line" padding>
 			<view class="example-body">
-				<uni-dateformat :date="now - 7200000"></uni-dateformat>
-				<uni-dateformat date="2020/12/12"></uni-dateformat>
+				<uni-dateformat :date="now - 7200000" locale="en"></uni-dateformat>
+				<uni-dateformat date="2020/12/12" locale="en"></uni-dateformat>
 			</view>
 		</uni-section>
 		<uni-section title="Set Threshold" subTitle="Threshold is used to control when to show just|immediately and when to show xx minutes ago|xx minutes later" type="line" padding>
 			<view class="example-body">
-				<uni-dateformat :date="now - 30000" :threshold="[0,3600000]"></uni-dateformat>
-				<uni-dateformat :date="now - 30000" :threshold="[0,0]"></uni-dateformat>
+				<uni-dateformat :date="now - 30000" :threshold="[0,3600000]" locale="en"></uni-dateformat>
+				<uni-dateformat :date="now - 30000" :threshold="[0,0]" locale="en"></uni-dateformat>
 			</view>
 		</uni-section>
 		<uni-section title="Use English" type="line" padding>